requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
ddtrace==0.44.0
decorator==5.0.9
dnspython==2.1.0
docker==4.3.1
fire==0.3.1
Flask==2.0.1
graphviz==0.14.1
grpcio==1.37.1
grpcio-tools==1.37.1
gunicorn==20.0.4
idna==2.10
intervaltree==3.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
loguru==0.5.3
lz4==3.1.0
MarkupSafe==2.0.1
mypy-protobuf==1.23
networkx==2.5
numpy==1.19.5
pandas==1.2.5
pinecone-client==0.8.59
protobuf==3.13.0
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4
pyzmq==19.0.2
requests==2.25.1
sentry-sdk==1.0.0
six==1.16.0
sortedcontainers==2.4.0
tenacity==7.0.0
termcolor==1.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==2.0.1
